Create initial Elm Sidebar Component
This commit creates an initial Elm application which renders a sidebar
component in place of the sidebar_template in formatter/html/templates.

The Elm sidebar at current retains all the stylings and class names of
the previous template but the functionality implement consists only of
opening and closing the sidebar, and selecting the category. Rendering
the nodes is not yet implemented as of this commit. Animations are also
not implemented.

The Elm project should be compiled and placed in
priv/ex_doc/formatter/html/assets/dist. This will be integrated into the build system in a future commit.