Integrate Elm into gulp build system.
Additionally, this commit adds code to the html formatter's footer
template which adds a functions field set to an empty array
to all module objects that ex_doc generates for modules which dont
contain any functions.
This is because the flags declared in Elm expect all modules
to have a functions field set to an array.